remove a type conversion warning to make buildable in MSVC

+++ aws-cpp-sdk-transfer-tests/TransferTests.cpp
@@ -104,7 +104,7 @@ protected:
         Aws::OFStream testFile;
         testFile.open(fileName.c_str());
         auto putStringLength = putString.length();
-        for (unsigned i = putStringLength; i <= fileSize; i += putStringLength)
+        for (size_t i = putStringLength; i <= fileSize; i += putStringLength)
         {
             testFile << putString;
         }
